Wrap errors so that they can be unwrapped later
Currently the source error's message is embedded into a new error but
the original error struct/type is discarded and unavailable to the
consumer of the client. By changing the %v to %w, the newly generated
error will wrap the original error so that the client can use
errors.Unwrap/.Is/.As to identify the original type of the error.